Mid-refactoring commit due to Claude usage limit reached.
